+++ versions/v.0.1.1/src/chunk/chunker.h
@@ -7,8 +7,12 @@
 enum class ChunkType {
     LOCATION,
     DEFINITION,
+    FUNCTION,
+    USAGE,
     HISTORY,
     TEMPORAL,
+    ADVANTAGES,
+    LIMITATIONS,
     PERSON,
     PROCEDURE,
     GENERAL